resolve: Move some diagnostic code into `diagnostics.rs`
